requirements.txt
llama-index-embeddings-azure-openai
llama-index-llms-azure-openai
llama-index-readers-file
llama-index
llama-index-core
openpyxl
python-dotenv